<==== Mercurial data ====>
Branch: vms

+++ vms/server/plugins/device/usb_camera_plugin/src/discovery_manager.cpp
@@ -2,7 +2,6 @@
 
 #include <QCryptographicHash>
 
-#include <nx/utils/app_info.h>
 #include <nx/utils/log/log.h>
 
 #include "plugin.h"
@@ -67,7 +66,7 @@ int DiscoveryManager::findCameras(nxcip::CameraInfo* cameras, const char* localI
     std::vector<DeviceDataWithNxId> devices = findCamerasInternal();
 
     int i;
-    for (i = 0; i < devices.size() && i < nxcip::CAMERA_INFO_ARRAY_SIZE; ++i)
+    for (i = 0; i < (int)devices.size() && i < nxcip::CAMERA_INFO_ARRAY_SIZE; ++i)
     {
         strncpy(
             cameras[i].modelName,
@@ -85,7 +84,7 @@ int DiscoveryManager::findCameras(nxcip::CameraInfo* cameras, const char* localI
 int DiscoveryManager::checkHostAddress(
     nxcip::CameraInfo* /*cameras*/,
     const char* /*address*/,
-    const char* /*login*/, 
+    const char* /*login*/,
     const char* /*password*/)
 {
     //host address doesn't mean anything for a local web cam
@@ -133,7 +132,7 @@ void DiscoveryManager::addOrUpdateCamera(const DeviceDataWithNxId& device)
     std::lock_guard<std::mutex> lock(m_mutex);
     auto it = m_cameras.find(device.nxId);
     NX_DEBUG(this, "addOrUpdateCamera attempting to add device: %1", device.toString());
-    if(it == m_cameras.end())
+    if (it == m_cameras.end())
     {
         NX_DEBUG(this, "Found new device: %1", device.toString());
         m_cameras.emplace(device.nxId, CameraAndDeviceDataWithNxId(device));
